+++ src/test/java/gdsc/konkuk/platformcore/controller/attendance/AttendanceControllerTest.java
@@ -4,8 +4,12 @@ import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import gdsc.konkuk.platformcore.application.attendance.AttendanceService;
+import gdsc.konkuk.platformcore.application.event.EventService;
+import gdsc.konkuk.platformcore.application.event.EventWithAttendance;
 import gdsc.konkuk.platformcore.domain.attendance.entity.Participant;
 
+import java.time.LocalDate;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -47,11 +51,10 @@ public class AttendanceControllerTest {
 
   private MockMvc mockMvc;
 
-  @Autowired
-  private ObjectMapper objectMapper;
+  @Autowired private ObjectMapper objectMapper;
 
-  @MockBean
-  private AttendanceService attendanceService;
+  @MockBean private AttendanceService attendanceService;
+  @MockBean private EventService eventService;
 
   @BeforeEach
   void setUp(
@@ -62,6 +65,51 @@ public class AttendanceControllerTest {
         .apply(springSecurity())
         .apply(documentationConfiguration(restDocumentation))
         .build();
+  }
+
+  @Test
+  @DisplayName("특정 달의 출석 정보를 조회할 수 있다")
+  @WithMockUser
+  void should_get_events_of_the_month_when_pass_year_month() throws Exception {
+    // given
+    given(eventService.getEventsOfTheMonthWithAttendance(any(LocalDate.class)))
+        .willReturn(
+            List.of(
+                // TODO: Fixture 정리
+                EventWithAttendance.builder().attendanceId(0L).eventId(0L).build(),
+                EventWithAttendance.builder().attendanceId(1L).eventId(1L).build(),
+                EventWithAttendance.builder().attendanceId(2L).eventId(2L).build()));
+
+    // when
+    ResultActions result =
+        mockMvc.perform(
+            RestDocumentationRequestBuilders.get("/api/v1/attendances?year=2021&month=1")
+                .with(csrf()));
+
+    // then
+    result
+        .andDo(print())
+        .andExpect(status().isOk())
+        .andDo(
+            document(
+                "eventsByMonth",
+                preprocessRequest(prettyPrint()),
+                preprocessResponse(prettyPrint()),
+                resource(
+                    ResourceSnippetParameters.builder()
+                        .description("특정 달의 출석 정보를 조회할 수 있다")
+                        .tag("attendance")
+                        .queryParameters(
+                            parameterWithName("year").description("년도"),
+                            parameterWithName("month").description("월"))
+                        .responseFields(
+                            fieldWithPath("success").description("성공 여부"),
+                            fieldWithPath("message").description("메시지"),
+                            fieldWithPath("data[].eventId").description("이벤트 ID"),
+                            fieldWithPath("data[].title").description("이벤트 제목"),
+                            fieldWithPath("data[].attendanceId").description("출석 ID"),
+                            fieldWithPath("data[].startAt").description("이벤트 시작 시간"))
+                        .build())));
   }
 
   @Test
