+++ src/test/java/gdsc/konkuk/platformcore/controller/event/EventControllerTest.java
@@ -4,7 +4,6 @@ import com.epages.restdocs.apispec.ResourceSnippetParameters;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gdsc.konkuk.platformcore.application.event.EventService;
-import gdsc.konkuk.platformcore.application.event.EventWithAttendance;
 
 import gdsc.konkuk.platformcore.domain.event.entity.Event;
 import gdsc.konkuk.platformcore.domain.event.entity.Retrospect;
@@ -25,9 +24,7 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 
 import static com.epages.restdocs.apispec.ResourceDocumentation.headerWithName;
 import static com.epages.restdocs.apispec.ResourceDocumentation.parameterWithName;
@@ -119,51 +116,6 @@ public class EventControllerTest {
                             fieldWithPath("success").description("성공 여부"),
                             fieldWithPath("message").description("메시지"),
                             fieldWithPath("data").description("null"))
-                        .build())));
-  }
-
-  @Test
-  @DisplayName("특정 달의 이벤트 정보를 조회할 수 있다")
-  @WithMockUser
-  void should_get_events_of_the_month_when_pass_year_month() throws Exception {
-    // given
-    given(eventService.getEventsOfTheMonthWithAttendance(any(LocalDate.class)))
-        .willReturn(
-            List.of(
-                // TODO: Fixture 정리
-                EventWithAttendance.builder().attendanceId(0L).eventId(0L).build(),
-                EventWithAttendance.builder().attendanceId(1L).eventId(1L).build(),
-                EventWithAttendance.builder().attendanceId(2L).eventId(2L).build()));
-
-    // when
-    ResultActions result =
-        mockMvc.perform(
-            RestDocumentationRequestBuilders.get("/api/v1/events?year=2021&month=1").with(csrf()));
-
-    // then
-    result
-        .andDo(print())
-        .andExpect(status().isOk())
-        .andDo(
-            document(
-                "eventsByMonth",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                resource(
-                    ResourceSnippetParameters.builder()
-                        .description("특정 달의 이벤트 정보를 조회할 수 있다")
-                        .tag("events")
-                        .queryParameters(
-                            parameterWithName("year").description("년도"),
-                            parameterWithName("month").description("월"))
-                        .responseFields(
-                            fieldWithPath("success").description("성공 여부"),
-                            fieldWithPath("message").description("메시지"),
-                            fieldWithPath("data[].eventId").description("이벤트 ID"),
-                            fieldWithPath("data[].title").description("이벤트 제목"),
-                            fieldWithPath("data[].attendanceId").description("출석 ID"),
-                            fieldWithPath("data[].thumbnailUrl").description("썸네일 URL"),
-                            fieldWithPath("data[].startAt").description("이벤트 시작 시간"))
                         .build())));
   }
